+++ resources/views/livewire/recruitment/components/staff-comments.blade.php
@@ -24,13 +24,14 @@
                                         <p>{{ $comment->body }}</p>
                                     </div>
                                     <div class="mt-2 text-sm space-x-2">
-                                        <span
-                                            class="text-gray-500 font-medium">{{ \Carbon\Carbon::parse($comment->created_at)->diffForHumans() }}</span>
-                                        <span class="text-gray-500 font-medium">&middot;</span>
-                                        <button type="button" wire:click="deleteComment('{{ $comment->uuid }}')"
-                                                class="text-gray-900 font-medium">
-                                            Delete
-                                        </button>
+                                        <span class="text-gray-500 font-medium">{{ \Carbon\Carbon::parse($comment->created_at)->diffForHumans() }}</span>
+                                        @can('delete', $comment)
+                                            <span class="text-gray-500 font-medium">&middot;</span>
+                                            <button type="button" wire:click="deleteComment('{{ $comment->uuid }}')"
+                                                    class="text-gray-900 font-medium">
+                                                Delete
+                                            </button>
+                                        @endcan
                                     </div>
                                 </div>
                             </div>
@@ -54,7 +55,7 @@
                         <div>
                             <label for="comment" class="sr-only">Comment</label>
                             <textarea id="comment" name="comment" rows="3" wire:model.lazy="comment"
-                                      class="shadow-sm block w-full focus:ring-blue-500 focus:border-blue-500 sm:text-sm border-gray-300 rounded-md @error('comment') border-red-300 text-red-900 placeholder-red-300 focus:border-red-300 focus:shadow-outline-red @enderror"
+                                      class="shadow-sm block w-full focus:ring-blue-500 focus:border-blue-500 sm:text-sm border-gray-300 rounded-md @error('comment') border-red-300 text-red-900 placeholder-red-300 focus:border-red-300 focus:ring-red @enderror"
                                       placeholder="Add a comment"></textarea>
                         </div>
                         @error('comment')
